pyproject.toml
[tool.poetry]
name = "slidegen"
version = "0.1.0"
description = ""
authors = ["Suman Das <59254445+sumandas0@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
dspy-ai = "^2.4.12"
fastapi = "^0.111.1"
google-generativeai = "^0.7.2"
gradio-client = "^1.1.0"
groq = "^0.9.0"
anthropic = "^0.31.2"
boto3 = "^1.34.145"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

